requirements.txt
Flask>=3.0
psycopg[binary]>=3.1
python-dotenv>=1.0
Werkzeug>=3.0
requests>=2.31
gunicorn>=21.0
redis>=4.5
psutil>=5.9
